+++ crates/network/src/any/mod.rs
@@ -1,10 +1,16 @@
-use alloy_rpc_types::{Extra, Header, Transaction, TransactionReceipt, TransactionRequest};
+use alloy_rpc_types::{
+    other::WithOtherFields, Header, Transaction, TransactionReceipt, TransactionRequest,
+};
 
 use crate::{Network, ReceiptResponse};
 
 mod builder;
 
 /// Types for a catch-all network.
+///
+/// Essentially just returns the regular Ethereum types + a catch all field.
+/// This [`Network`] should be used only when the network is not known at
+/// compile time.
 #[derive(Debug, Clone, Copy)]
 pub struct AnyNetwork {
     _private: (),
@@ -19,16 +25,16 @@ impl Network for AnyNetwork {
 
     type Header = alloy_consensus::Header;
 
-    type TransactionRequest = Extra<TransactionRequest>;
+    type TransactionRequest = WithOtherFields<TransactionRequest>;
 
-    type TransactionResponse = Extra<Transaction>;
+    type TransactionResponse = WithOtherFields<Transaction>;
 
-    type ReceiptResponse = Extra<TransactionReceipt>;
+    type ReceiptResponse = WithOtherFields<TransactionReceipt>;
 
-    type HeaderResponse = Extra<Header>;
+    type HeaderResponse = WithOtherFields<Header>;
 }
 
-impl ReceiptResponse for Extra<TransactionReceipt> {
+impl ReceiptResponse for WithOtherFields<TransactionReceipt> {
     fn contract_address(&self) -> Option<alloy_primitives::Address> {
         self.contract_address
     }
